requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
torch==1.9.1
torchvision==0.10.1
transformers==4.11.3
Pillow==8.4.0
